Splitted up the profilepage component. Still need to split the editing stuff to separate component after Firebase db functionality is implemented.
